Added Tag and Comment CRUD operations; Created the Article view page, added commenting and rating features; Added user and tag follow DB schema; Added DB schema constraints; Added aggregate rating and view count columns;

+++ Backend/SkillForge/SkillForge/Models/Database/ArticleRating.cs
@@ -10,5 +10,5 @@ public class ArticleRating : BaseEntity
 
     public int ArticleId { get; set; }
 
-    public byte Rate { get; set; }
+    public short Rate { get; set; }
 }
